#62:	IMPROVEMENT: Because of BUG#61: Generate for every generator of a modul and the modules itself used for generate sample-sound playing a buffer hirarchy. 		The buffer contains the samples and a integrated signal (maybe also a differetial) over time.

+++ NoiseComp/src/de/schmiereck/noiseComp/soundSource/SoundSamplesBufferData.java
@@ -40,12 +40,12 @@ public class SoundSamplesBufferData
 		synchronized (this)
 		{
 System.out.println("clearBuffer: " + startTimePos + ", " + endTimePos);
-			long startFrame = (int)(startTimePos * this.frameRate);
-			long endFrame = (int)(endTimePos * this.frameRate);
+			int startFrame = (int)(startTimePos * this.frameRate);
+			int endFrame = (int)(endTimePos * this.frameRate);
 	
-			for (long pos = startFrame; pos < endFrame; pos++)
+			for (int pos = startFrame; pos < endFrame; pos++)
 			{
-				this.bufferSoundSamples[(int)pos] = null;
+				this.bufferSoundSamples[pos] = null;
 			}
 			
 			this.emptyBufferStart 	= Math.min(this.emptyBufferStart, startFrame);
